+++ docs/CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [3.0.4] - 2022-11-26
 ### Fixed
 - [Issue 235](https://github.com/aza547/wow-recorder/issues/235) - Fix a bug where abandoned M+ runs caused the app to crash. 
+- [NO-ISSUE] - Fix a bug where closing WoW while in an activity could crash the app.
 
 ## [3.0.3] - 2022-11-24
 ### Fixed